integrate upscaling
