+++ server/app.py
@@ -1,18 +1,21 @@
-from flask import Flask, request, jsonify, send_from_directory
-from werkzeug.utils import secure_filename
+import base64
+
+from flask import Flask, request, jsonify
 import os
+import io
 from PIL import Image
 import logging
 from flask_cors import CORS
-from utils import success_response, error_response
-import uuid
+
+from super_resolution.execute import upscale_gan
+from super_resolution.utils import error_response
 
 
 app = Flask(__name__)
 
 CORS(app, resources={
     r"/*": {
-        "origins": ["http://localhost:5173"],
+        "origins": ["http://localhost:5173", "http://127.0.0.1:5173"],
         "methods": ["GET", "POST", "OPTIONS"],
         "allow_headers": ["Content-Type"]
     }
@@ -21,11 +24,15 @@ CORS(app, resources={
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-UPLOAD_FOLDER = 'uploads'
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg'}
 MAX_CONTENT_LENGTH = 16 * 1024 * 1024
 
-app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+FORMAT_MAP = {
+    'jpg': ('JPEG', 'image/jpeg'),
+    'jpeg': ('JPEG', 'image/jpeg'),
+    'png': ('PNG', 'image/png')
+}
+
 
 @app.route('/', methods=['GET'])
 def home():
@@ -34,60 +41,13 @@ def home():
         'message': 'Server is running'
     })
 
-def allowed_file(filename):
-    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
+def allowed_file(filename, ext):
+    return ext in ALLOWED_EXTENSIONS
+
 
 def get_file_size(file_path):
     return os.path.getsize(file_path)
 
-
-@app.route('/upload', methods=['OPTIONS'])
-def upload_options():
-    response = app.make_default_options_response()
-    response.headers['Access-Control-Allow-Origin'] = 'http://localhost:5173'
-    response.headers['Access-Control-Allow-Methods'] = 'POST, OPTIONS'
-    response.headers['Access-Control-Allow-Headers'] = 'Content-Type'
-    return response
-
-
-@app.route('/upload', methods=['POST'])
-def upload_image():
-    try:
-        file = request.files['file']
-        if file:
-            img = Image.open(file)
-            upscale_factor = 2
-            img_resized = img.resize((img.width * upscale_factor, img.height * upscale_factor))
-            filename = secure_filename(file.filename)
-            img_resized.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
-            return jsonify({'status': 'success', 'url': f'/uploads/{filename}'})
-    except Exception as e:
-        print(f"Error processing upload: {e}")
-        return jsonify({'status': 'error', 'message': str(e)})
-
-#Upscaling logic (replace with an actual implementation later)
-def process_image(file_path):
-    with Image.open(file_path) as img:
-        mock_upscale_factor = 2.0
-        original_width, original_height = img.size
-        new_width = int(original_width * mock_upscale_factor)
-        new_height = int(original_height * mock_upscale_factor)
-        
-        upscaled_img = img.resize((new_width, new_height), Image.LANCZOS)
-
-        new_filename = f"transformed_{uuid.uuid4().hex}.png"
-        upscaled_img_path = os.path.join(app.config['UPLOAD_FOLDER'], new_filename)
-        upscaled_img.save(upscaled_img_path)
-
-        print(f"Original size: {original_width}x{original_height}")
-        print(f"Transformed size: {new_width}x{new_height}")
-
-        return {
-            'success': True,
-            'upscaled_image_path': upscaled_img_path,
-            'new_filename': new_filename,
-            'mock_upscale_factor': mock_upscale_factor
-        }
         
 @app.route('/upscale', methods=['POST'])
 def upscale_image():
@@ -96,28 +56,29 @@ def upscale_image():
             return error_response('No file selected or invalid file in request', 400)
 
         file = request.files['file']
+        filename = file.filename
+        ext = filename.rsplit('.', 1)[1].lower()
+        pil_format, mime_type = FORMAT_MAP.get(ext, ('JPEG', 'image/jpeg'))
 
-        if not allowed_file(file.filename):
+        if not allowed_file(filename, ext):
             return error_response('Invalid file type. Allowed types: PNG, JPG, JPEG', 400)
 
-        filename = secure_filename(file.filename)
-        os.makedirs(app.config['UPLOAD_FOLDER'], exist_ok=True)
-        file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        file.save(file_path)
+        upscaled_array = upscale_gan(file)
+        upscaled_image = Image.fromarray(upscaled_array[:, :, ::-1])
 
-        #apply upscaling
-        processing_result = process_image(file_path)
+        img_io = io.BytesIO()
+        upscaled_image.save(img_io, format=pil_format)  # or PNG
+        img_io.seek(0)
+        img_bytes = img_io.read()
 
-        if not processing_result['success']:
-            return error_response('Error processing image', 500, processing_result['error'])
-
-        upscaled_image_url = f"http://localhost:5100/uploads/{processing_result['new_filename']}"
+        # Base64 encode the image bytes
+        img_b64 = base64.b64encode(img_bytes).decode('utf-8')
 
         return jsonify({
             'status': 'success',
             'message': 'Image upscaled successfully',
-            'url': upscaled_image_url,
-            'mock_upscale_factor': 2.0
+            'image': img_b64,
+            'format': ext
         })
 
     except Exception as e:
@@ -127,11 +88,7 @@ def upscale_image():
             'message': 'Error processing upscale',
             'details': str(e)
         }), 500
-        
-        
-@app.route('/uploads/<filename>')
-def uploaded_file(filename):
-    return send_from_directory(app.config['UPLOAD_FOLDER'], filename)
+
     
 if __name__ == '__main__':
     app.run(debug=True, port=5100)
